Implement 16-bit PNG save/load funcs to pass test
The test passes, but it isn't a true unit test; it's more of an
integration test, since it doesn't check the intermediate result, i.e.
the result of saving the 16-bit PNG to file.

In other words, the 16-bit PNG save function needs its own unit tests,
at the very least to make sure that the produced file is a valid 16-bit
PNG file.  On top of that, somehow the tests musts verify that the saved
16-bit PNG file contains the expected pixel data.

Similarly, the 16-bit PNG load function needs its own unit tests, which
need to verify that if fed any 16-bit PNG image known to be valid, the
load function correctly loads the image into memory.  To verify that the
image gets loaded correctly, maybe the test needs to display the PNG to
make sure the image looks as expected, but then that wouldn't be an
automated test since it would require human eyes/human judgment to tell
whether the test passes.  Not sure how to address this yet..