Add test and non-test
This commit adds a test to verify that an exception gets thrown if an
attempt is made to load an 8-bit PNG file using the load_png_16_file()
function.

This commit also adds a non-test that simply writes a 16-bit PNG file to
disk without deleting it.  Currently, to verify that the output file is
a valid 16-bit PNG file, it must be done manually using external tools,
e.g. imagemagick's "identify" utility.  I'm not sure yet how to automate
this.