Add 16-bit png load-save-load test
This test passes and appears to show that the 16-bit png load and save
functions are working, but I'm not sure yet, because the copy of the png
file has a different size than the original png file. The copy is 12 KB,
while the original is less than 4 KB. However, the copy is viewable in
third party viewing software such as GNOME Image Viewer, and visually
the copy appears identical to the original.  The size discrepancy might
just be due to the compression factor being used in save_png_16_file(),
which seems to be set by the call to png_set_compression_level()...